added dynamic urls, add and delete item in those dynamic url lists

+++ app.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require('mongoose');
+const _ = require("lodash");
+
 const app = express();
 app.set('view engine', 'ejs');
 
@@ -12,7 +14,7 @@ app.use(express.static("public"));
 mongoose.connect("mongodb://localhost:27017/todolistDB", {
   useNewUrlParser: true,
   useUnifiedTopology: true,
-  useFindAndModify: false 
+  useFindAndModify: false
 });
 const itemsSchema = {
   name: String,
@@ -33,6 +35,11 @@ const item3 = new Item({
   check: "off"
 });
 
+const listSchema = {
+  name: String,
+  items: [itemsSchema]
+}
+const List = mongoose.model("List", listSchema);
 
 let today = new Date();
 let options = {
@@ -45,6 +52,14 @@ let workItems = [];
 
 app.get("/", (req, res) => {
   Item.find({}, (err, foundItems) => {
+    List.find({}, (err, found) => {
+      if (found.length === 0) {
+        const list = new List({
+          name: "Initializer"
+        });
+        list.save();
+      }
+    });
     if (foundItems.length === 0) {
       Item.insertMany([item1, item2, item3], (err) => {
         if (err) console.log(err);
@@ -62,12 +77,23 @@ app.get("/", (req, res) => {
 
 app.post("/", function(req, res) {
   const itemName = req.body.newItem;
+  const listName = req.body.list;
   const item = new Item({
     name: itemName,
     check: "off"
   });
-  item.save();
-  res.redirect("/");
+  if (listName === date) {
+    item.save();
+    res.redirect("/");
+  } else {
+    List.findOne({
+      name: listName
+    }, (err, foundList) => {
+      foundList.items.push(item);
+      foundList.save();
+      res.redirect("/" + listName);
+    })
+  }
 });
 
 app.post("/check", (req, res) => {
@@ -75,16 +101,19 @@ app.post("/check", (req, res) => {
   Item.findById(checkItemId, (err, item) => {
     if (err) console.log(err);
     else {
-      // console.log(item.check);
       if (item.check == "off") {
-        Item.findByIdAndUpdate({_id: checkItemId}, {
+        Item.findByIdAndUpdate({
+          _id: checkItemId
+        }, {
           check: "on"
         }, (err) => {
           if (err) console.log(err);
           else console.log("Successfully checked");
         });
       } else if (item.check == "on") {
-        Item.findByIdAndUpdate({_id: checkItemId}, {
+        Item.findByIdAndUpdate({
+          _id: checkItemId
+        }, {
           check: "off"
         }, (err) => {
           if (err) console.log(err);
@@ -98,19 +127,56 @@ app.post("/check", (req, res) => {
 
 app.post("/delete", (req, res) => {
   const deletedItemId = req.body.bin;
-  Item.findByIdAndRemove(deletedItemId, (err) => {
-    if (err) console.log(err);
-    else console.log("Successfully deleted item from database.");
-  });
-  res.redirect("/");
+  const listName = req.body.listName;
+  if (listName === date) {
+    Item.findByIdAndRemove(deletedItemId, (err) => {
+      if (err) console.log(err);
+      else console.log("Successfully deleted item from database.");
+    });
+    res.redirect("/");
+  } else {
+    List.findOneAndUpdate({
+      name: listName
+    }, {
+      $pull: {
+        items: {
+          _id: deletedItemId
+        }
+      }
+    }, (err, foundList) => {
+      if (!err) res.redirect("/" + listName);
+    })
+  }
+
 });
 
-app.get("/work", (req, res) => {
-  res.render("list", {
-    listTitle: "Work List",
-    newListItems: workItems
+app.get("/favicon.ico", (req, res) => {
+  return 1;
+});
+
+app.get("/:customListName", (req, res) => {
+  const customListName = _.capitalize(req.params.customListName);
+  List.findOne({
+    name: customListName
+  }, (err, foundList) => {
+    if (!err) {
+      if (!foundList) {
+        const list = new List({
+          name: customListName,
+          items: [item1, item2, item3]
+        });
+        list.save();
+        res.redirect("/" + customListName);
+      } else {
+        res.render("list", {
+          listTitle: foundList.name,
+          newListItems: foundList.items
+        });
+      }
+    }
   });
-})
+
+});
 
 app.post("/work", (req, res) => {
   let item = req.body.newItem;
